click outside to dismiss expanded
