+++ Composer/packages/form-dialogs/src/components/property/FormDialogSchemaDetails.tsx
@@ -3,7 +3,6 @@
 
 import styled from '@emotion/styled';
 import { NeutralColors } from '@uifabric/fluent-theme';
-import { Stack } from 'office-ui-fabric-react/lib/Stack';
 import * as React from 'react';
 import { DragDropContext, DropResult } from 'react-beautiful-dnd';
 import { useRecoilValue } from 'recoil';
@@ -11,6 +10,15 @@ import { formDialogSchemaAtom } from 'src/atoms/appState';
 import { useHandlers } from 'src/atoms/handlers';
 import { PropertyRequiredKind } from 'src/atoms/types';
 import { PropertyList } from 'src/components/property/PropertyList';
+import { Lifetime } from 'src/utils/base';
+import { jsPropertyListClassName } from 'src/utils/constants';
+
+const Root = styled.div({
+  display: 'flex',
+  flexDirection: 'column',
+  height: '100%',
+  width: '100%',
+});
 
 const Separator = styled.div({
   background: NeutralColors.gray60,
@@ -21,10 +29,47 @@ const Separator = styled.div({
 
 export const FormDialogSchemaDetails = () => {
   const { optionalPropertyIds, requiredPropertyIds } = useRecoilValue(formDialogSchemaAtom);
-  const { moveProperty } = useHandlers();
+  const { moveProperty, activatePropertyId } = useHandlers();
+  const containerRef = React.useRef<HTMLDivElement>();
+  const dragEventRef = React.useRef<boolean>(false);
+
+  React.useEffect(() => {
+    const lifetime = new Lifetime();
+
+    const clickOutsideLists = (e: MouseEvent) => {
+      const { x, y } = e;
+      const elms = Array.prototype.slice.call(
+        containerRef.current.querySelectorAll(`.${jsPropertyListClassName}`) || []
+      ) as HTMLDivElement[];
+      if (
+        !dragEventRef.current &&
+        !elms.some((elm) => {
+          const { left, right, top, bottom } = elm.getBoundingClientRect();
+          return x <= right && x >= left && y <= bottom && y >= top;
+        })
+      ) {
+        activatePropertyId({ id: '' });
+      }
+    };
+
+    if (containerRef.current) {
+      containerRef.current.addEventListener('click', clickOutsideLists);
+      lifetime.add(() => {
+        containerRef.current.removeEventListener('click', clickOutsideLists);
+      });
+    }
+
+    return () => lifetime.dispose();
+  }, []);
+
+  const dragStart = React.useCallback(() => {
+    dragEventRef.current = true;
+  }, []);
 
   const dragEnd = React.useCallback(
     (result: DropResult) => {
+      dragEventRef.current = false;
+
       const { source, destination, draggableId } = result;
 
       if (!destination || (destination.droppableId === source.droppableId && destination.index === source.index)) {
@@ -43,12 +88,12 @@ export const FormDialogSchemaDetails = () => {
   );
 
   return (
-    <Stack grow verticalFill>
-      <DragDropContext onDragEnd={dragEnd}>
+    <Root ref={containerRef}>
+      <DragDropContext onDragEnd={dragEnd} onDragStart={dragStart}>
         <PropertyList kind="required" propertyIds={requiredPropertyIds} />
         <Separator />
         <PropertyList kind="optional" propertyIds={optionalPropertyIds} />
       </DragDropContext>
-    </Stack>
+    </Root>
   );
 };
